+++ gensym/ulid.go
@@ -1,30 +1,30 @@
-package server
+package gensym
 
 import (
 	"context"
-	"errors"
 	"math/rand"
-	"strings"
 	"sync"
 
 	"github.com/deref/exo/chrono"
-	ulid "github.com/oklog/ulid/v2"
+	"github.com/oklog/ulid/v2"
 )
 
-type idGen struct {
+// A concurrency-safe source of monotonic ULIDs.
+// See <https://github.com/ulid/spec>.
+type ULIDGenerator struct {
 	mu      sync.Mutex
 	entropy *ulid.MonotonicEntropy
 }
 
-func newIdGen(ctx context.Context) *idGen {
+func NewULIDGenerator(ctx context.Context) *ULIDGenerator {
 	seed := int64(chrono.NowNano(ctx))
 	randRead := rand.New(rand.NewSource(seed))
-	return &idGen{
+	return &ULIDGenerator{
 		entropy: ulid.Monotonic(randRead, 0),
 	}
 }
 
-func (gen *idGen) nextId(ctx context.Context) ([]byte, error) {
+func (gen *ULIDGenerator) NextID(ctx context.Context) ([]byte, error) {
 	gen.mu.Lock()
 	defer gen.mu.Unlock()
 	ts := ulid.Timestamp(chrono.Now(ctx))
@@ -33,13 +33,4 @@ func (gen *idGen) nextId(ctx context.Context) ([]byte, error) {
 		return nil, err
 	}
 	return id.MarshalBinary()
-}
-
-func parseID(id []byte) (string, error) {
-	var asULID ulid.ULID
-	if copy(asULID[:], id) != 16 {
-		return "", errors.New("invalid length")
-	}
-
-	return strings.ToLower(asULID.String()), nil
 }
